Refactor to functional component to fix #20 and split rendering filters to `<DataTableToolbarFilterRenderFilters />`
